fixing case of FileNames
