+++ src/main/java/com/inkysea/vmware/vra/jenkins/plugin/VRABuildWrapper.java
@@ -21,7 +21,7 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
-//import java.util.concurrent.TimeoutException;
+//import java.util.concurrent.TimeoutException; 
 
 import org.kohsuke.stapler.DataBoundConstructor;
