+++ src/main/java/com/inkysea/vmware/vra/jenkins/plugin/VRADeploymentDestroyPostBuildAction.java
@@ -1,10 +1,11 @@
 /*
  * To change this template, choose Tools | Templates
- * and open the template in the editor.
+ * and open the template in the editor. 
  */
 package com.inkysea.vmware.vra.jenkins.plugin;
 
 import com.inkysea.vmware.vra.jenkins.plugin.model.Deployment;
+import com.inkysea.vmware.vra.jenkins.plugin.model.DestroyParam;
 import com.inkysea.vmware.vra.jenkins.plugin.model.PluginParam;
 import com.inkysea.vmware.vra.jenkins.plugin.model.RequestParam;
 import com.inkysea.vmware.vra.jenkins.plugin.util.EnvVariableResolver;
@@ -19,26 +20,25 @@ import java.io.IOException;
 import java.io.PrintStream;
 import java.util.ArrayList;
 import java.util.Collection;
-import java.util.Collections;
 import java.util.List;
 import java.util.logging.Logger;
 
 
-public class VRADeploymentPostBuildAction extends Notifier {
+public class VRADeploymentDestroyPostBuildAction extends Notifier {
 
-	private static final Logger LOGGER = Logger.getLogger(VRADeploymentPostBuildAction.class.getName());
+	private static final Logger LOGGER = Logger.getLogger(VRADeploymentDestroyPostBuildAction.class.getName());
 
-	protected List<PluginParam> params;
+	protected List<DestroyParam> destroyParams;
 	protected List<Deployment> deployments = new ArrayList<Deployment>();
-	private List<RequestParam> requestParams;
 
 	@DataBoundConstructor
-	public VRADeploymentPostBuildAction(List<PluginParam> params) {
-		this.params = params;
+	public VRADeploymentDestroyPostBuildAction(List<DestroyParam> destroyParams) {
+		this.destroyParams = destroyParams;
 	}
 
-	public List<PluginParam> getParams() {
-		return params;
+	// method name must match with the property field in config.jelly
+	public List<DestroyParam> getDestroyParams() {
+		return destroyParams;
 	}
 
 	public BuildStepMonitor getRequiredMonitorService() {
@@ -76,36 +76,23 @@ public class VRADeploymentPostBuildAction extends Notifier {
 		boolean success = true;
 
 		int counter = 1;
-		for (PluginParam param : params) {
-
-			// Resolve any build variables included in the request paramaters.
-			List<RequestParam> rparamResolved = new ArrayList<RequestParam>();;
-
-			for(RequestParam rparam : param.getRequestParams()){
-				String rparamString =  helper.replaceBuildParamWithValue(rparam.getRequestParam().toString());
-				rparamResolved.add(new RequestParam(rparamString));
-			}
+		for (DestroyParam param : destroyParams) {
 
 			// Resolve any environment variables in the parameters
-			PluginParam fparam = new PluginParam(helper.replaceBuildParamWithValue(param.getServerUrl()),
+			DestroyParam fparam = new DestroyParam(helper.replaceBuildParamWithValue(param.getServerUrl()),
 					helper.replaceBuildParamWithValue(param.getUserName()),
 					helper.replaceBuildParamWithValue(param.getPassword()),
 					helper.replaceBuildParamWithValue(param.getTenant()),
-					helper.replaceBuildParamWithValue(param.getBlueprintName()),
-					param.isWaitExec(),param.getRequestTemplate(), rparamResolved);
+					helper.replaceBuildParamWithValue(param.getDeploymentName()));
 
-			final Deployment deployment = newDeployment(listener.getLogger(), fparam);
-
-
-			if (deployment.create()) {
+			final Deployment deployment = new Deployment(listener.getLogger(), fparam);
+			if(deployment.destroy(env.expand(param.getDeploymentName()))) {
 				this.deployments.add(deployment);
 
-				//change counter to string and append pb for build environment
-				String strCounter = "PB_"+Integer.toString(counter);
-				env.putAll(deployment.getDeploymentComponents(strCounter));
+				LOGGER.info("Success");
 				counter++;
 			} else {
-				build.setResult(Result.FAILURE);
+				LOGGER.warning("Failed");
 				success = false;
 				break;
 			}
@@ -116,29 +103,20 @@ public class VRADeploymentPostBuildAction extends Notifier {
 	}
 
 
-	protected Deployment newDeployment(PrintStream logger, PluginParam params) throws IOException {
-
-		Boolean isURL = false;
-		String recipe = null;
-
-		return new Deployment(logger, params);
-
-	}
-
 	@Override
 	public BuildStepDescriptor getDescriptor() {
 		return DESCRIPTOR;
 	}
 
 	@Extension
-	public static final VRADeploymentPostBuildAction.DescriptorImpl DESCRIPTOR = new VRADeploymentPostBuildAction.DescriptorImpl();
+	public static final DescriptorImpl DESCRIPTOR = new DescriptorImpl();
 
 	public static class DescriptorImpl extends BuildStepDescriptor<Publisher> {
 
 		@Override
 		public String getDisplayName() {
                     
-			return "vRealize Automation Deployment";
+			return "Destroy vRealize Automation Deployment";
 		}
 
 		@Override
