+++ src/main/resources/com/inkysea/vmware/vra/jenkins/plugin/VRABlueprintBuildStep/config.jelly
@@ -1,0 +1,10 @@
+<?jelly escape-by-default='true'?>
+<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define"
+         xmlns:l="/lib/layout" xmlns:t="/lib/hudson" xmlns:f="/lib/form">
+
+	<f:entry>
+		<f:repeatableProperty field="params" minimum="1" header="Blueprint configuration" add="Add vRA Blueprint">
+		</f:repeatableProperty>
+	</f:entry>
+
+</j:jelly>
