+++ src/main/resources/com/inkysea/vmware/vra/jenkins/plugin/VRABlueprintBuildStep/help.jelly
@@ -1,0 +1,4 @@
+<?jelly escape-by-default='true'?>
+<div>
+Loads a vRA Blueprint as part of the build steps.
+</div>
