+++ src/main/resources/com/inkysea/vmware/vra/jenkins/plugin/VRABuildWrapper/help.jelly
@@ -1,0 +1,4 @@
+<?jelly escape-by-default='true'?>
+<div>
+Provisions a vRA deployment environment prior to the build and then deletes the deployment at build completion.
+</div>
