+++ src/main/resources/com/inkysea/vmware/vra/jenkins/plugin/VRADeploymentDestroyPostBuildAction/config.jelly
@@ -1,0 +1,11 @@
+<?jelly escape-by-default='true'?>
+<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define"
+         xmlns:l="/lib/layout" xmlns:t="/lib/hudson" xmlns:f="/lib/form">
+
+	<f:entry>
+		<f:repeatableProperty field="destroyParams" minimum="1" header="Deployment configuration" add="Destroy another vRA deployment">
+		</f:repeatableProperty>
+	</f:entry>
+
+
+</j:jelly>
