+++ src/main/resources/com/inkysea/vmware/vra/jenkins/plugin/VRADeploymentDestroyPostBuildAction/help.jelly
@@ -1,0 +1,4 @@
+<?jelly escape-by-default='true'?>
+<div>
+Provisions a vRA deployment environment as part of the build steps.
+</div>
