+++ src/main/resources/com/inkysea/vmware/vra/jenkins/plugin/VRANotifier/config.jelly
@@ -1,0 +1,59 @@
+<?jelly escape-by-default='true'?>
+<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define"
+         xmlns:l="/lib/layout" xmlns:t="/lib/hudson" xmlns:f="/lib/form">
+
+        <f:entry title="vRealize Automation URL"  field="serverUrl">
+            <f:textbox name="serverUrl" />
+        </f:entry>
+
+        <f:entry title="Tenant"  field="tenant">
+             <f:textbox name="tenant" />
+        </f:entry>
+
+        <f:entry title="User Name"  field="userName">
+             <f:textbox name="userName" />
+        </f:entry>
+
+        <f:entry title="Password"  field="password">
+             <f:password name="password" />
+        </f:entry>
+
+        <f:entry title="Blueprint"  field="blueprintName">
+             <f:textbox name="blueprintName" />
+        </f:entry>
+
+        <f:entry title="Execute and Wait"  field="waitExec">
+             <f:checkbox default="true" name="waitExec" />
+        </f:entry>
+
+        <f:entry title="Destroy Deployment on job completion"  field="destroyDeployment">
+                     <f:checkbox default="true"  name="destroyDeployment" />
+         </f:entry>
+
+       <f:entry>
+             <f:repeatable name="deploymentParms" field="deploymentParms" noAddButton="false" add="Add Parameter">
+                 <table width="100%">
+                      <f:entry title="Parameter Name">
+                         <f:textbox field="name" />
+                         </f:entry>
+                         <f:entry title="Parameter Value">
+                         <f:textbox field="value" />
+                         </f:entry>
+                         <f:entry>
+                         <div align="right" class="repeatable-delete show-if-only" style="margin-left: 1em;">
+                             <f:repeatableDeleteButton value="Delete Parameter" /><br/>
+                            <!-- <input type="button" value="${%Add Parameter}" class="repeatable-add show-if-last" />  -->
+                         </div>
+
+                         </f:entry>
+                 </table>
+             </f:repeatable>
+       </f:entry>
+
+        <f:entry title="">
+                 <div align="right">
+                   <f:repeatableDeleteButton />
+                 </div>
+        </f:entry>
+
+</j:jelly>
