+++ src/main/resources/com/inkysea/vmware/vra/jenkins/plugin/VRANotifier/help-bluePrint.html
@@ -1,0 +1,3 @@
+<div>
+   Blue Print
+</div>
