+++ src/main/resources/com/inkysea/vmware/vra/jenkins/plugin/VRANotifier/help-password.html
@@ -1,3 +1,3 @@
 <div>
-   Blueprint
+    Password
 </div>
