+++ src/main/resources/com/inkysea/vmware/vra/jenkins/plugin/VRANotifier/help-tenant.html
@@ -1,3 +1,3 @@
 <div>
-   Blueprint
+  Tenant
 </div>
