+++ src/main/resources/com/inkysea/vmware/vra/jenkins/plugin/VRANotifier/help-userName.html
@@ -1,3 +1,3 @@
 <div>
-   Blueprint
+   User name
 </div>
