+++ src/main/resources/com/inkysea/vmware/vra/jenkins/plugin/VRANotifier/help-vraUrl.html
@@ -1,0 +1,3 @@
+<div>
+  The URL for the vRealize Automation host.  ex:  https://vra.domain.com
+</div>
